chore: moving course detail related code to Discovery framwrok
