+++ Course/Course/Data/Persistence/CourseCoreModel.xcdatamodeld/CourseCoreModel.xcdatamodel/contents
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="22222" systemVersion="22G91" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="22225" systemVersion="23C64" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="CDCourseBlock" representedClassName="CDCourseBlock" syncable="YES" codeGenerationType="class">
         <attribute name="allSources" optional="YES" attributeType="Transformable" valueTransformerName="NSSecureUnarchiveFromData" customClassName="[String]"/>
         <attribute name="blockId" optional="YES" attributeType="String"/>
@@ -43,24 +43,6 @@
         <attribute name="missedDeadlines" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
         <attribute name="missedGatedContent" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
         <attribute name="verifiedUpgradeLink" optional="YES" attributeType="String"/>
-    </entity>
-    <entity name="CDCourseDetails" representedClassName=".CDCourseDetails" syncable="YES" codeGenerationType="class">
-        <attribute name="courseBannerURL" attributeType="String"/>
-        <attribute name="courseDescription" attributeType="String"/>
-        <attribute name="courseEnd" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
-        <attribute name="courseID" attributeType="String"/>
-        <attribute name="courseStart" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
-        <attribute name="courseTitle" attributeType="String"/>
-        <attribute name="enrollmentEnd" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
-        <attribute name="enrollmentStart" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
-        <attribute name="isEnrolled" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
-        <attribute name="org" attributeType="String"/>
-        <attribute name="overviewHTML" attributeType="String"/>
-        <uniquenessConstraints>
-            <uniquenessConstraint>
-                <constraint value="courseID"/>
-            </uniquenessConstraint>
-        </uniquenessConstraints>
     </entity>
     <entity name="CDCourseItem" representedClassName=".CDCourseItem" syncable="YES" codeGenerationType="class">
         <attribute name="certificate" optional="YES" attributeType="String"/>
