+++ Course/Course/Presentation/CourseAnalytics.swift
@@ -9,9 +9,6 @@ import Foundation
 
 //sourcery: AutoMockable
 public protocol CourseAnalytics {
-    func courseEnrollClicked(courseId: String, courseName: String)
-    func courseEnrollSuccess(courseId: String, courseName: String)
-    func viewCourseClicked(courseId: String, courseName: String)
     func resumeCourseTapped(courseId: String, courseName: String, blockId: String)
     func sequentialClicked(courseId: String, courseName: String, blockId: String, blockName: String)
     func verticalClicked(courseId: String, courseName: String, blockId: String, blockName: String)
@@ -29,9 +26,6 @@ public protocol CourseAnalytics {
 
 #if DEBUG
 class CourseAnalyticsMock: CourseAnalytics {
-    public func courseEnrollClicked(courseId: String, courseName: String) {}
-    public func courseEnrollSuccess(courseId: String, courseName: String) {}
-    public func viewCourseClicked(courseId: String, courseName: String) {}
     public func resumeCourseTapped(courseId: String, courseName: String, blockId: String) {}
     public func sequentialClicked(courseId: String, courseName: String, blockId: String, blockName: String) {}
     public func verticalClicked(courseId: String, courseName: String, blockId: String, blockName: String) {}
