+++ Course/Course/en.lproj/Localizable.strings
@@ -6,11 +6,6 @@
   
 */
 
-"DETAILS.TITLE" = "Course details";
-"DETAILS.VIEW_COURSE" = "View course";
-"DETAILS.ENROLL_NOW" = "Enroll now";
-"DETAILS.ENROLLMENT_DATE_IS_OVER" = "You cannot enroll in this course because the enrollment date is over.";
-
 "OUTLINE.CONGRATULATIONS" = "Congratulations!";
 "OUTLINE.PASSED_THE_COURSE" = "You've passed the course";
 "OUTLINE.VIEW_CERTIFICATE" = "View certificate";
