+++ Discovery/Discovery/Data/Model/CourseDetails.swift
