+++ Discovery/Discovery/Data/Persistence/DiscoveryCoreModel.xcdatamodeld/DiscoveryCoreModel.xcdatamodel/contents
@@ -1,5 +1,23 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="21754" systemVersion="22F82" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="22225" systemVersion="23C64" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+    <entity name="CDCourseDetails" representedClassName=".CDCourseDetails" syncable="YES" codeGenerationType="class">
+        <attribute name="courseBannerURL" attributeType="String"/>
+        <attribute name="courseDescription" attributeType="String"/>
+        <attribute name="courseEnd" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
+        <attribute name="courseID" attributeType="String"/>
+        <attribute name="courseStart" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
+        <attribute name="courseTitle" attributeType="String"/>
+        <attribute name="enrollmentEnd" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
+        <attribute name="enrollmentStart" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
+        <attribute name="isEnrolled" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
+        <attribute name="org" attributeType="String"/>
+        <attribute name="overviewHTML" attributeType="String"/>
+        <uniquenessConstraints>
+            <uniquenessConstraint>
+                <constraint value="courseID"/>
+            </uniquenessConstraint>
+        </uniquenessConstraints>
+    </entity>
     <entity name="CDDiscoveryCourse" representedClassName=".CDDiscoveryCourse" syncable="YES" codeGenerationType="class">
         <attribute name="certificate" optional="YES" attributeType="String"/>
         <attribute name="courseCount" optional="YES" attributeType="Integer 32" defaultValueString="0" usesScalarValueType="YES"/>
