+++ Discovery/Discovery/Data/Persistence/DiscoveryPersistenceProtocol.swift
@@ -11,6 +11,8 @@ import Core
 public protocol DiscoveryPersistenceProtocol {
     func loadDiscovery() throws -> [CourseItem]
     func saveDiscovery(items: [CourseItem])
+    func loadCourseDetails(courseID: String) throws -> CourseDetails
+    func saveCourseDetails(course: CourseDetails)
 }
 
 public final class DiscoveryBundle {
