+++ Discovery/Discovery/Presentation/DiscoveryAnalytics.swift
@@ -12,6 +12,9 @@ public protocol DiscoveryAnalytics {
     func discoverySearchBarClicked()
     func discoveryCoursesSearch(label: String, coursesCount: Int)
     func discoveryCourseClicked(courseID: String, courseName: String)
+    func viewCourseClicked(courseId: String, courseName: String)
+    func courseEnrollClicked(courseId: String, courseName: String)
+    func courseEnrollSuccess(courseId: String, courseName: String)
 }
 
 #if DEBUG
@@ -19,5 +22,8 @@ class DiscoveryAnalyticsMock: DiscoveryAnalytics {
     public func discoverySearchBarClicked() {}
     public func discoveryCoursesSearch(label: String, coursesCount: Int) {}
     public func discoveryCourseClicked(courseID: String, courseName: String) {}
+    public func viewCourseClicked(courseId: String, courseName: String) {}
+    public func courseEnrollClicked(courseId: String, courseName: String) {}
+    public func courseEnrollSuccess(courseId: String, courseName: String) {}
 }
 #endif
