+++ OpenEdX/DI/ScreenAssembly.swift
@@ -113,7 +113,7 @@ class ScreenAssembly: Assembly {
             DiscoveryWebviewModel(
                 router: r.resolve(DiscoveryRouter.self)!,
                 config: r.resolve(ConfigProtocol.self)!,
-                interactor: r.resolve(CourseInteractorProtocol.self)!,
+                interactor: r.resolve(DiscoveryInteractorProtocol.self)!,
                 connectivity: r.resolve(ConnectivityProtocol.self)!,
                 analytics: r.resolve(DiscoveryAnalytics.self)!,
                 storage: r.resolve(CoreStorage.self)!,
@@ -227,9 +227,9 @@ class ScreenAssembly: Assembly {
         }
         container.register(CourseDetailsViewModel.self) { r in
             CourseDetailsViewModel(
-                interactor: r.resolve(CourseInteractorProtocol.self)!,
-                router: r.resolve(CourseRouter.self)!,
-                analytics: r.resolve(CourseAnalytics.self)!,
+                interactor: r.resolve(DiscoveryInteractorProtocol.self)!,
+                router: r.resolve(DiscoveryRouter.self)!,
+                analytics: r.resolve(DiscoveryAnalytics.self)!,
                 config: r.resolve(ConfigProtocol.self)!,
                 cssInjector: r.resolve(CSSInjector.self)!,
                 connectivity: r.resolve(ConnectivityProtocol.self)!,
