+++ OpenEdX/Data/CoursePersistence.swift
@@ -18,47 +18,6 @@ public class CoursePersistence: CoursePersistenceProtocol {
         self.context = context
     }
     
-    public func loadCourseDetails(courseID: String) throws -> CourseDetails {
-        let request = CDCourseDetails.fetchRequest()
-        request.predicate = NSPredicate(format: "courseID = %@", courseID)
-        guard let courseDetails = try? context.fetch(request).first else { throw NoCachedDataError() }
-        return CourseDetails(courseID: courseDetails.courseID ?? "",
-                             org: courseDetails.org ?? "",
-                             courseTitle: courseDetails.courseTitle ?? "",
-                             courseDescription: courseDetails.courseDescription ?? "",
-                             courseStart: courseDetails.courseStart,
-                             courseEnd: courseDetails.courseEnd,
-                             enrollmentStart: courseDetails.enrollmentStart,
-                             enrollmentEnd: courseDetails.enrollmentEnd,
-                             isEnrolled: courseDetails.isEnrolled,
-                             overviewHTML: courseDetails.overviewHTML ?? "",
-                             courseBannerURL: courseDetails.courseBannerURL ?? "",
-                             courseVideoURL: nil)
-    }
-    
-    public func saveCourseDetails(course: CourseDetails) {
-        context.performAndWait {
-            let newCourseDetails = CDCourseDetails(context: self.context)
-            newCourseDetails.courseID = course.courseID
-            newCourseDetails.org = course.org
-            newCourseDetails.courseTitle = course.courseTitle
-            newCourseDetails.courseDescription = course.courseDescription
-            newCourseDetails.courseStart = course.courseStart
-            newCourseDetails.courseEnd = course.courseEnd
-            newCourseDetails.enrollmentStart = course.enrollmentStart
-            newCourseDetails.enrollmentEnd = course.enrollmentEnd
-            newCourseDetails.isEnrolled = course.isEnrolled
-            newCourseDetails.overviewHTML = course.overviewHTML
-            newCourseDetails.courseBannerURL = course.courseBannerURL
-            
-            do {
-                try context.save()
-            } catch {
-                print("⛔️⛔️⛔️⛔️⛔️", error)
-            }
-        }
-    }
-    
     public func loadEnrollments() throws -> [CourseItem] {
         let result = try? context.fetch(CDCourseItem.fetchRequest())
             .map {
